feat(Input): formatCurrencyをformatNumericStringにリネームして公開する
CurrencyInput内部専用だった数値フォーマット関数は通貨に限定されない汎用的な処理のため、
公開APIとしての名前をformatNumericStringに変更し、汎用ユーティリティとしてsrc/libsに移動した。
