+++ packages/smarthr-ui/src/components/Input/CurrencyInput/CurrencyInput.tsx
@@ -13,9 +13,8 @@ import {
 import { useLatest } from '../../../hooks/useLatest'
 import { useMergeRefs } from '../../../hooks/useMergeRefs'
 import { useOnce } from '../../../hooks/useOnce'
+import { formatNumericString } from '../../../libs/formatNumericString'
 import { Input } from '../Input'
-
-import { formatCurrency } from './currencyInputHelper'
 
 type Props = Omit<ComponentProps<typeof Input>, 'type' | 'value' | 'defaultValue'> & {
   /** 通貨の値 */
@@ -47,7 +46,7 @@ export const CurrencyInput = forwardRef<HTMLInputElement, Props>(
         }
       }
       const formatCurrencyValue = (raw = '') => {
-        formatValue(formatCurrency(raw))
+        formatValue(formatNumericString(raw))
       }
 
       return {
