+++ packages/smarthr-ui/src/components/Input/index.ts
@@ -1,3 +1,4 @@
 export { Input, backgroundColor } from './Input'
 export { CurrencyInput } from './CurrencyInput'
+export { formatNumericString } from '../../libs/formatNumericString'
 export { SearchInput } from './SearchInput'
